cleanup exeption structure and move class defns to util
