+++ algobattle/docker_util.py
@@ -17,7 +17,7 @@ from pydantic import Field
 from anyio.to_thread import run_sync
 from urllib3.exceptions import ReadTimeoutError
 
-from algobattle.util import Encodable, Role, TempDir, inherit_docs, BaseModel
+from algobattle.util import AlgobattleBaseException, BuildError, DockerError, Encodable, EncodingError, ExceptionInfo, ExecutionError, ExecutionTimeout, Role, TempDir, inherit_docs, BaseModel
 from algobattle.problem import Problem
 
 
@@ -58,61 +58,6 @@ def client() -> DockerClient:
 def get_os_type() -> Literal["linux", "windows"]:
     """OS running inside docker containers."""
     return client().info()["OSType"]
-
-
-class _BaseException(Exception):
-    def __init__(self, message: str, detail: str | None = None, *args: object) -> None:
-        self.message = message
-        self.detail = detail
-        super().__init__(*args)
-
-
-class BuildError(_BaseException):
-    """Indicates that the build process could not be completed successfully."""
-
-
-class ProgramError(_BaseException):
-    """Parent class for exceptions raised during the execution of a program."""
-
-
-class ExecutionError(ProgramError):
-    """Indicates that the program could not be executed successfully."""
-
-    def __init__(self, message: str, detail: str = "", *args: object, runtime: float) -> None:
-        self.runtime = runtime
-        super().__init__(message, detail, *args)
-
-
-class ExecutionTimeout(ExecutionError):
-    """Indicates that the program ran into the timeout."""
-
-
-class EncodingError(ProgramError):
-    """Indicates that the given data could not be encoded or decoded properly."""
-
-    def __init__(self, message: str = "", detail: str | None = None, *args: object) -> None:
-        super().__init__(message, detail, *args)
-
-
-class DockerError(ProgramError, BuildError):
-    """Indicates that an issue with the docker daemon occured."""
-
-
-class ExceptionInfo(BaseModel):
-    """An exception that can be encoded into a json file."""
-
-    type: str
-    message: str
-    detail: str | None = None
-
-    @classmethod
-    def from_exception(cls, error: _BaseException) -> Self:
-        """Constructs an instance from a raised exception."""
-        return cls(
-            type=error.__class__.__name__,
-            message=error.message,
-            detail=error.detail,
-        )
 
 
 class ProgramUiProxy(Protocol):
@@ -232,9 +177,9 @@ class Image:
         except Timeout as e:
             raise BuildError(f"Build process for '{image_name}' ran into a timeout.") from e
         except DockerBuildError as e:
-            raise BuildError(f"Building '{image_name}' did not complete successfully.", e.msg) from e
+            raise BuildError(f"Building '{image_name}' did not complete successfully.", detail=e.msg) from e
         except APIError as e:
-            raise DockerError(f"Docker APIError thrown while building '{image_name}'.", str(e)) from e
+            raise DockerError(f"Docker APIError thrown while building '{image_name}'.", detail=str(e)) from e
 
         return cls(image_name, cast(str, image.id), description if description is not None else image_name, path=path)
 
@@ -317,13 +262,13 @@ class Image:
         except ImageNotFound as e:
             raise RuntimeError(f"Image {self.name} (id={self.id}) does not exist") from e
         except APIError as e:
-            raise DockerError(f"Docker APIError thrown while running '{self.name}'.", str(e)) from e
+            raise DockerError(f"Docker APIError thrown while running '{self.name}'.", detail=str(e)) from e
         finally:
             if container is not None:
                 try:
                     container.remove(force=True)
                 except APIError as e:
-                    raise DockerError(f"Couldn't remove {name}", str(e)) from e
+                    raise DockerError(f"Couldn't remove {name}", detail=str(e)) from e
 
         return elapsed_time
 
@@ -356,7 +301,7 @@ class Image:
                     file.write(chunk)
             image.remove(force=True)
         except APIError as e:
-            raise DockerError(f"Docker APIError thrown while archiving '{self.name}'", str(e)) from e
+            raise DockerError(f"Docker APIError thrown while archiving '{self.name}'", detail=str(e)) from e
         return ArchivedImage(path, self.name, self.id, self.description)
 
     def _run_container(self, container: DockerContainer, timeout: float | None = None) -> float:
@@ -371,7 +316,7 @@ class Image:
             else:
                 raise ExecutionError(
                     "The program executed in the container crashed.",
-                    f"exit code: {response['StatusCode']}, error message:\n{container.logs().decode()}",
+                    detail=f"exit code: {response['StatusCode']}, error message:\n{container.logs().decode()}",
                     runtime=elapsed_time,
                 )
         except (Timeout, ConnectionError) as e:
@@ -485,7 +430,7 @@ class Program(ABC):
         with TempDir() as input, TempDir() as output:
             try:
                 self._setup_folders(input, output, size, input_instance)
-            except ProgramError as e:
+            except AlgobattleBaseException as e:
                 return result_class(ProgramRunInfo(params=run_params, runtime=0, error=ExceptionInfo.from_exception(e)))
             with open(input / "info.json", "w+") as f:
                 json.dump(
@@ -522,7 +467,7 @@ class Program(ABC):
                         error=ExceptionInfo.from_exception(e),
                     )
                 )
-            except ProgramError as e:
+            except AlgobattleBaseException as e:
                 return result_class(
                     ProgramRunInfo(
                         params=run_params,
@@ -533,7 +478,7 @@ class Program(ABC):
 
             try:
                 output_data = self._parse_output(output, size, input_instance)
-            except ProgramError as e:
+            except AlgobattleBaseException as e:
                 return result_class(
                     ProgramRunInfo(
                         params=run_params,
@@ -608,7 +553,7 @@ class Generator(Program):
             except EncodingError:
                 raise
             except Exception as e:
-                raise EncodingError("Error thrown while decoding the solution.", str(e)) from e
+                raise EncodingError("Error thrown while decoding the solution.", detail=str(e)) from e
             if not solution.is_valid(problem, size):
                 raise EncodingError("Solution is not valid.")
         else:
